feat: remove connection delegate
