--- src/main/java/io/vertx/core/http/impl/Http2StreamImpl.java
@@ -1,72 +1,0 @@
-package io.vertx.core.http.impl;
-
-import io.netty.handler.codec.http2.Http2Exception;
-import io.netty.handler.codec.http2.Http2Headers;
-import io.netty.handler.codec.http2.Http2Stream;
-import io.vertx.core.AsyncResult;
-import io.vertx.core.Future;
-import io.vertx.core.Handler;
-import io.vertx.core.http.HttpVersion;
-import io.vertx.core.http.impl.headers.VertxDefaultHttp2Headers;
-import io.vertx.core.http.impl.headers.VertxDefaultHttpHeaders;
-import io.vertx.core.impl.ContextInternal;
-import io.vertx.core.spi.metrics.ClientMetrics;
-import io.vertx.core.tracing.TracingPolicy;
-
-class Http2StreamImpl extends HttpStreamImpl<Http2ClientConnection, Http2Stream, Http2Headers> {
-  Http2StreamImpl(Http2ClientConnection conn, ContextInternal context, boolean push,
-                  VertxHttpConnectionDelegate<Http2Stream, Http2Headers> connectionDelegate, ClientMetrics metrics) {
-    super(conn, context, push, connectionDelegate, metrics);
-  }
-
-  @Override
-  public HttpClientConnection connection() {
-    return conn;
-  }
-
-  @Override
-  public HttpVersion version() {
-    return HttpVersion.HTTP_2;
-  }
-
-  @Override
-  void metricsEnd(HttpStream<?, ?, ?> stream) {
-    conn.metricsEnd(stream);
-  }
-
-  @Override
-  void recycle() {
-    conn.recycle();
-  }
-
-  @Override
-  int lastStreamCreated() {
-    return this.conn.handler.encoder().connection().local().lastStreamCreated();
-  }
-
-  @Override
-  protected void createStreamInternal(int id, boolean b, Handler<AsyncResult<Http2Stream>> onComplete) throws HttpException {
-    try {
-      Http2Stream stream = this.conn.handler.encoder().connection().local().createStream(id, false);
-      onComplete.handle(Future.succeededFuture(stream));
-    } catch (Http2Exception e) {
-      throw new HttpException(e);
-    }
-  }
-
-  @Override
-  protected TracingPolicy getTracingPolicy() {
-    return conn.client.options().getTracingPolicy();
-  }
-
-  @Override
-  protected boolean isTryUseCompression() {
-    return this.conn.client.options().isTryUseCompression();
-  }
-
-  @Override
-  VertxDefaultHttpHeaders<Http2Headers> createHttpHeadersWrapper() {
-    return new VertxDefaultHttp2Headers();
-  }
-
-}
