--- src/main/java/io/vertx/core/http/impl/VertxHttpConnectionDelegate.java
@@ -1,45 +1,0 @@
-package io.vertx.core.http.impl;
-
-import io.netty.buffer.ByteBuf;
-import io.netty.handler.codec.Headers;
-import io.netty.util.concurrent.FutureListener;
-import io.vertx.core.MultiMap;
-import io.vertx.core.http.StreamPriority;
-
-interface VertxHttpConnectionDelegate<S, H extends Headers<CharSequence, CharSequence, H>> {
-  void consumeCredits(int len);
-
-  void writeFrame(byte type, short flags, ByteBuf payload);
-
-  void writeHeaders(H headers, boolean end, int dependency, short weight, boolean exclusive, boolean checkFlush,
-                    FutureListener<Void> promise);
-
-  void writePriorityFrame(StreamPriority priority);
-
-  void writeData(ByteBuf chunk, boolean end, FutureListener<Void> promise);
-
-  void writeReset(int streamId, long code);
-
-  void init(VertxHttpStreamBase vertxHttpStream, S stream);
-
-  int getStreamId();
-
-  boolean remoteSideOpen();
-
-  boolean hasStream();
-
-  MultiMap getEmptyHeaders();
-
-  boolean isWritable();
-
-  boolean isTrailersReceived();
-
-  long getWindowSize();
-
-  CharSequence getHeaderMethod(H headers);
-
-  String getHeaderStatus(H headers);
-
-  MultiMap createHeaderAdapter(H headers);
-
-}
